test: split compileScript macro tests
