+++ packages/compiler-sfc/__tests__/compileScript.spec.ts
@@ -68,82 +68,6 @@ describe('SFC compile <script setup>', () => {
     assertCode(content)
   })
 
-  test('defineProps()', () => {
-    const { content, bindings } = compile(`
-<script setup>
-const props = defineProps({
-  foo: String
-})
-const bar = 1
-</script>
-  `)
-    // should generate working code
-    assertCode(content)
-    // should analyze bindings
-    expect(bindings).toStrictEqual({
-      foo: BindingTypes.PROPS,
-      bar: BindingTypes.LITERAL_CONST,
-      props: BindingTypes.SETUP_REACTIVE_CONST
-    })
-
-    // should remove defineOptions import and call
-    expect(content).not.toMatch('defineProps')
-    // should generate correct setup signature
-    expect(content).toMatch(`setup(__props, { expose: __expose }) {`)
-    // should assign user identifier to it
-    expect(content).toMatch(`const props = __props`)
-    // should include context options in default export
-    expect(content).toMatch(`export default {
-  props: {
-  foo: String
-},`)
-  })
-
-  test('defineProps w/ external definition', () => {
-    const { content } = compile(`
-    <script setup>
-    import { propsModel } from './props'
-    const props = defineProps(propsModel)
-    </script>
-      `)
-    assertCode(content)
-    expect(content).toMatch(`export default {
-  props: propsModel,`)
-  })
-
-  // #4764
-  test('defineProps w/ leading code', () => {
-    const { content } = compile(`
-    <script setup>import { x } from './x'
-    const props = defineProps({})
-    </script>
-    `)
-    // props declaration should be inside setup, not moved along with the import
-    expect(content).not.toMatch(`const props = __props\nimport`)
-    assertCode(content)
-  })
-
-  test('defineEmits()', () => {
-    const { content, bindings } = compile(`
-<script setup>
-const myEmit = defineEmits(['foo', 'bar'])
-</script>
-  `)
-    assertCode(content)
-    expect(bindings).toStrictEqual({
-      myEmit: BindingTypes.SETUP_CONST
-    })
-    // should remove defineEmits import and call
-    expect(content).not.toMatch('defineEmits')
-    // should generate correct setup signature
-    expect(content).toMatch(
-      `setup(__props, { expose: __expose, emit: myEmit }) {`
-    )
-    // should include context options in default export
-    expect(content).toMatch(`export default {
-  emits: ['foo', 'bar'],`)
-  })
-
   test('defineProps/defineEmits in multi-variable declaration', () => {
     const { content } = compile(`
     <script setup>
@@ -200,282 +124,6 @@ const myEmit = defineEmits(['foo', 'bar'])
     assertCode(content)
     expect(content).toMatch(`props: ['item'],`)
     expect(content).toMatch(`emits: ['a'],`)
-  })
-
-  describe('defineOptions()', () => {
-    test('basic usage', () => {
-      const { content } = compile(`
-        <script setup>
-        defineOptions({ name: 'FooApp' })
-        </script>
-      `)
-      assertCode(content)
-      // should remove defineOptions import and call
-      expect(content).not.toMatch('defineOptions')
-      // should include context options in default export
-      expect(content).toMatch(
-        `export default /*#__PURE__*/Object.assign({ name: 'FooApp' }, `
-      )
-    })
-
-    test('empty argument', () => {
-      const { content } = compile(`
-        <script setup>
-        defineOptions()
-        </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`export default {`)
-      // should remove defineOptions import and call
-      expect(content).not.toMatch('defineOptions')
-    })
-
-    it('should emit an error with two defineProps', () => {
-      expect(() =>
-        compile(`
-        <script setup>
-        defineOptions({ name: 'FooApp' })
-        defineOptions({ name: 'BarApp' })
-        </script>
-        `)
-      ).toThrowError('[@vue/compiler-sfc] duplicate defineOptions() call')
-    })
-
-    it('should emit an error with props or emits property', () => {
-      expect(() =>
-        compile(`
-        <script setup>
-        defineOptions({ props: { foo: String } })
-        </script>
-        `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare props. Use defineProps() instead.'
-      )
-
-      expect(() =>
-        compile(`
-        <script setup>
-        defineOptions({ emits: ['update'] })
-        </script>
-      `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare emits. Use defineEmits() instead.'
-      )
-
-      expect(() =>
-        compile(`
-        <script setup>
-        defineOptions({ expose: ['foo'] })
-        </script>
-      `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare expose. Use defineExpose() instead.'
-      )
-
-      expect(() =>
-        compile(`
-        <script setup>
-        defineOptions({ slots: ['foo'] })
-        </script>
-      `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare slots. Use defineSlots() instead.'
-      )
-    })
-
-    it('should emit an error with type generic', () => {
-      expect(() =>
-        compile(`
-        <script setup lang="ts">
-        defineOptions<{ name: 'FooApp' }>()
-        </script>
-        `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot accept type arguments'
-      )
-    })
-
-    it('should emit an error with type assertion', () => {
-      expect(() =>
-        compile(`
-        <script setup lang="ts">
-        defineOptions({ props: [] } as any)
-        </script>
-        `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare props. Use defineProps() instead.'
-      )
-    })
-
-    it('should emit an error with declaring props/emits/slots/expose', () => {
-      expect(() =>
-        compile(`
-          <script setup>
-          defineOptions({ props: ['foo'] })
-          </script>
-        `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare props. Use defineProps() instead'
-      )
-
-      expect(() =>
-        compile(`
-          <script setup>
-          defineOptions({ emits: ['update'] })
-          </script>
-        `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare emits. Use defineEmits() instead'
-      )
-
-      expect(() =>
-        compile(`
-          <script setup>
-          defineOptions({ expose: ['foo'] })
-          </script>
-        `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare expose. Use defineExpose() instead'
-      )
-
-      expect(() =>
-        compile(`
-          <script setup lang="ts">
-          defineOptions({ slots: Object })
-          </script>
-        `)
-      ).toThrowError(
-        '[@vue/compiler-sfc] defineOptions() cannot be used to declare slots. Use defineSlots() instead'
-      )
-    })
-  })
-
-  test('defineExpose()', () => {
-    const { content } = compile(`
-<script setup>
-defineExpose({ foo: 123 })
-</script>
-  `)
-    assertCode(content)
-    // should remove defineOptions import and call
-    expect(content).not.toMatch('defineExpose')
-    // should generate correct setup signature
-    expect(content).toMatch(`setup(__props, { expose: __expose }) {`)
-    // should replace callee
-    expect(content).toMatch(/\b__expose\(\{ foo: 123 \}\)/)
-  })
-
-  describe('defineModel()', () => {
-    test('basic usage', () => {
-      const { content, bindings } = compile(
-        `
-        <script setup>
-        const modelValue = defineModel({ required: true })
-        const c = defineModel('count')
-        </script>
-        `,
-        { defineModel: true }
-      )
-      assertCode(content)
-      expect(content).toMatch('props: {')
-      expect(content).toMatch('"modelValue": { required: true },')
-      expect(content).toMatch('"count": {},')
-      expect(content).toMatch('emits: ["update:modelValue", "update:count"],')
-      expect(content).toMatch(
-        `const modelValue = _useModel(__props, "modelValue")`
-      )
-      expect(content).toMatch(`const c = _useModel(__props, "count")`)
-      expect(content).toMatch(`return { modelValue, c }`)
-      expect(content).not.toMatch('defineModel')
-
-      expect(bindings).toStrictEqual({
-        modelValue: BindingTypes.SETUP_REF,
-        count: BindingTypes.PROPS,
-        c: BindingTypes.SETUP_REF
-      })
-    })
-
-    test('w/ defineProps and defineEmits', () => {
-      const { content, bindings } = compile(
-        `
-        <script setup>
-        defineProps({ foo: String })
-        defineEmits(['change'])
-        const count = defineModel({ default: 0 })
-        </script>
-      `,
-        { defineModel: true }
-      )
-      assertCode(content)
-      expect(content).toMatch(`props: _mergeModels({ foo: String }`)
-      expect(content).toMatch(`"modelValue": { default: 0 }`)
-      expect(content).toMatch(`const count = _useModel(__props, "modelValue")`)
-      expect(content).not.toMatch('defineModel')
-      expect(bindings).toStrictEqual({
-        count: BindingTypes.SETUP_REF,
-        foo: BindingTypes.PROPS,
-        modelValue: BindingTypes.PROPS
-      })
-    })
-
-    test('w/ array props', () => {
-      const { content, bindings } = compile(
-        `
-        <script setup>
-        defineProps(['foo', 'bar'])
-        const count = defineModel('count')
-        </script>
-      `,
-        { defineModel: true }
-      )
-      assertCode(content)
-      expect(content).toMatch(`props: _mergeModels(['foo', 'bar'], {
-    "count": {},
-  })`)
-      expect(content).toMatch(`const count = _useModel(__props, "count")`)
-      expect(content).not.toMatch('defineModel')
-      expect(bindings).toStrictEqual({
-        foo: BindingTypes.PROPS,
-        bar: BindingTypes.PROPS,
-        count: BindingTypes.SETUP_REF
-      })
-    })
-
-    test('w/ local flag', () => {
-      const { content } = compile(
-        `<script setup>
-        const foo = defineModel({ local: true, default: 1 })
-        const bar = defineModel('bar', { [key]: true })
-        const baz = defineModel('baz', { ...x })
-        const qux = defineModel('qux', x)
-
-        const foo2 = defineModel('foo2', { local: true, ...x })
-
-        const local = true
-        const hoist = defineModel('hoist', { local })
-        </script>`,
-        { defineModel: true }
-      )
-      assertCode(content)
-      expect(content).toMatch(
-        `_useModel(__props, "modelValue", { local: true })`
-      )
-      expect(content).toMatch(`_useModel(__props, "bar", { [key]: true })`)
-      expect(content).toMatch(`_useModel(__props, "baz", { ...x })`)
-      expect(content).toMatch(`_useModel(__props, "qux", x)`)
-      expect(content).toMatch(`_useModel(__props, "foo2", { local: true })`)
-      expect(content).toMatch(`_useModel(__props, "hoist", { local })`)
-    })
-  })
-
-  test('<script> after <script setup> the script content not end with `\\n`', () => {
-    const { content } = compile(`
-    <script setup>
-    import { x } from './x'
-    </script>
-    <script>const n = 1</script>
-    `)
-    assertCode(content)
   })
 
   describe('<script> and <script setup> co-usage', () => {
@@ -1168,739 +816,6 @@ defineExpose({ foo: 123 })
       assertCode(content)
     })
 
-    test('defineProps/Emit w/ runtime options', () => {
-      const { content } = compile(`
-<script setup lang="ts">
-const props = defineProps({ foo: String })
-const emit = defineEmits(['a', 'b'])
-</script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`export default /*#__PURE__*/_defineComponent({
-  props: { foo: String },
-  emits: ['a', 'b'],
-  setup(__props, { expose: __expose, emit }) {`)
-    })
-
-    test('defineProps w/ type', () => {
-      const { content, bindings } = compile(`
-      <script setup lang="ts">
-      interface Test {}
-
-      type Alias = number[]
-
-      defineProps<{
-        string: string
-        number: number
-        boolean: boolean
-        object: object
-        objectLiteral: { a: number }
-        fn: (n: number) => void
-        functionRef: Function
-        objectRef: Object
-        dateTime: Date
-        array: string[]
-        arrayRef: Array<any>
-        tuple: [number, number]
-        set: Set<string>
-        literal: 'foo'
-        optional?: any
-        recordRef: Record<string, null>
-        interface: Test
-        alias: Alias
-        method(): void
-        symbol: symbol
-        extract: Extract<1 | 2 | boolean, 2>
-        exclude: Exclude<1 | 2 | boolean, 2>
-        uppercase: Uppercase<'foo'>
-        params: Parameters<(foo: any) => void>
-        nonNull: NonNullable<string | null>
-        objectOrFn: {
-          (): void
-          foo: string
-        }
-
-        union: string | number
-        literalUnion: 'foo' | 'bar'
-        literalUnionNumber: 1 | 2 | 3 | 4 | 5
-        literalUnionMixed: 'foo' | 1 | boolean
-        intersection: Test & {}
-        intersection2: 'foo' & ('foo' | 'bar')
-        foo: ((item: any) => boolean) | null
-
-        unknown: UnknownType
-        unknownUnion: UnknownType | string
-        unknownIntersection: UnknownType & Object
-        unknownUnionWithBoolean: UnknownType | boolean
-        unknownUnionWithFunction: UnknownType | (() => any)
-      }>()
-      </script>`)
-      assertCode(content)
-      expect(content).toMatch(`string: { type: String, required: true }`)
-      expect(content).toMatch(`number: { type: Number, required: true }`)
-      expect(content).toMatch(`boolean: { type: Boolean, required: true }`)
-      expect(content).toMatch(`object: { type: Object, required: true }`)
-      expect(content).toMatch(`objectLiteral: { type: Object, required: true }`)
-      expect(content).toMatch(`fn: { type: Function, required: true }`)
-      expect(content).toMatch(`functionRef: { type: Function, required: true }`)
-      expect(content).toMatch(`objectRef: { type: Object, required: true }`)
-      expect(content).toMatch(`dateTime: { type: Date, required: true }`)
-      expect(content).toMatch(`array: { type: Array, required: true }`)
-      expect(content).toMatch(`arrayRef: { type: Array, required: true }`)
-      expect(content).toMatch(`tuple: { type: Array, required: true }`)
-      expect(content).toMatch(`set: { type: Set, required: true }`)
-      expect(content).toMatch(`literal: { type: String, required: true }`)
-      expect(content).toMatch(`optional: { type: null, required: false }`)
-      expect(content).toMatch(`recordRef: { type: Object, required: true }`)
-      expect(content).toMatch(`interface: { type: Object, required: true }`)
-      expect(content).toMatch(`alias: { type: Array, required: true }`)
-      expect(content).toMatch(`method: { type: Function, required: true }`)
-      expect(content).toMatch(`symbol: { type: Symbol, required: true }`)
-      expect(content).toMatch(
-        `objectOrFn: { type: [Function, Object], required: true },`
-      )
-      expect(content).toMatch(`extract: { type: Number, required: true }`)
-      expect(content).toMatch(
-        `exclude: { type: [Number, Boolean], required: true }`
-      )
-      expect(content).toMatch(`uppercase: { type: String, required: true }`)
-      expect(content).toMatch(`params: { type: Array, required: true }`)
-      expect(content).toMatch(`nonNull: { type: String, required: true }`)
-      expect(content).toMatch(
-        `union: { type: [String, Number], required: true }`
-      )
-      expect(content).toMatch(`literalUnion: { type: String, required: true }`)
-      expect(content).toMatch(
-        `literalUnionNumber: { type: Number, required: true }`
-      )
-      expect(content).toMatch(
-        `literalUnionMixed: { type: [String, Number, Boolean], required: true }`
-      )
-      expect(content).toMatch(`intersection: { type: Object, required: true }`)
-      expect(content).toMatch(`intersection2: { type: String, required: true }`)
-      expect(content).toMatch(`foo: { type: [Function, null], required: true }`)
-      expect(content).toMatch(`unknown: { type: null, required: true }`)
-      // uninon containing unknown type: skip check
-      expect(content).toMatch(`unknownUnion: { type: null, required: true }`)
-      // intersection containing unknown type: narrow to the known types
-      expect(content).toMatch(
-        `unknownIntersection: { type: Object, required: true },`
-      )
-      expect(content).toMatch(
-        `unknownUnionWithBoolean: { type: Boolean, required: true, skipCheck: true },`
-      )
-      expect(content).toMatch(
-        `unknownUnionWithFunction: { type: Function, required: true, skipCheck: true }`
-      )
-      expect(bindings).toStrictEqual({
-        string: BindingTypes.PROPS,
-        number: BindingTypes.PROPS,
-        boolean: BindingTypes.PROPS,
-        object: BindingTypes.PROPS,
-        objectLiteral: BindingTypes.PROPS,
-        fn: BindingTypes.PROPS,
-        functionRef: BindingTypes.PROPS,
-        objectRef: BindingTypes.PROPS,
-        dateTime: BindingTypes.PROPS,
-        array: BindingTypes.PROPS,
-        arrayRef: BindingTypes.PROPS,
-        tuple: BindingTypes.PROPS,
-        set: BindingTypes.PROPS,
-        literal: BindingTypes.PROPS,
-        optional: BindingTypes.PROPS,
-        recordRef: BindingTypes.PROPS,
-        interface: BindingTypes.PROPS,
-        alias: BindingTypes.PROPS,
-        method: BindingTypes.PROPS,
-        symbol: BindingTypes.PROPS,
-        objectOrFn: BindingTypes.PROPS,
-        extract: BindingTypes.PROPS,
-        exclude: BindingTypes.PROPS,
-        union: BindingTypes.PROPS,
-        literalUnion: BindingTypes.PROPS,
-        literalUnionNumber: BindingTypes.PROPS,
-        literalUnionMixed: BindingTypes.PROPS,
-        intersection: BindingTypes.PROPS,
-        intersection2: BindingTypes.PROPS,
-        foo: BindingTypes.PROPS,
-        uppercase: BindingTypes.PROPS,
-        params: BindingTypes.PROPS,
-        nonNull: BindingTypes.PROPS,
-        unknown: BindingTypes.PROPS,
-        unknownUnion: BindingTypes.PROPS,
-        unknownIntersection: BindingTypes.PROPS,
-        unknownUnionWithBoolean: BindingTypes.PROPS,
-        unknownUnionWithFunction: BindingTypes.PROPS
-      })
-    })
-
-    test('defineProps w/ interface', () => {
-      const { content, bindings } = compile(`
-      <script setup lang="ts">
-      interface Props { x?: number }
-      defineProps<Props>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`x: { type: Number, required: false }`)
-      expect(bindings).toStrictEqual({
-        x: BindingTypes.PROPS
-      })
-    })
-
-    test('defineProps w/ extends interface', () => {
-      const { content, bindings } = compile(`
-      <script lang="ts">
-        interface Foo { x?: number }
-      </script>
-      <script setup lang="ts">
-        interface Bar extends Foo { y?: number }
-        interface Props extends Bar {
-          z: number
-          y: string
-        }
-        defineProps<Props>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`z: { type: Number, required: true }`)
-      expect(content).toMatch(`y: { type: String, required: true }`)
-      expect(content).toMatch(`x: { type: Number, required: false }`)
-      expect(bindings).toStrictEqual({
-        x: BindingTypes.PROPS,
-        y: BindingTypes.PROPS,
-        z: BindingTypes.PROPS
-      })
-    })
-
-    test('defineProps w/ exported interface', () => {
-      const { content, bindings } = compile(`
-      <script setup lang="ts">
-      export interface Props { x?: number }
-      defineProps<Props>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`x: { type: Number, required: false }`)
-      expect(bindings).toStrictEqual({
-        x: BindingTypes.PROPS
-      })
-    })
-
-    test('defineProps w/ exported interface in normal script', () => {
-      const { content, bindings } = compile(`
-      <script lang="ts">
-        export interface Props { x?: number }
-      </script>
-      <script setup lang="ts">
-        defineProps<Props>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`x: { type: Number, required: false }`)
-      expect(bindings).toStrictEqual({
-        x: BindingTypes.PROPS
-      })
-    })
-
-    test('defineProps w/ type alias', () => {
-      const { content, bindings } = compile(`
-      <script setup lang="ts">
-      type Props = { x?: number }
-      defineProps<Props>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`x: { type: Number, required: false }`)
-      expect(bindings).toStrictEqual({
-        x: BindingTypes.PROPS
-      })
-    })
-
-    test('defineProps w/ exported type alias', () => {
-      const { content, bindings } = compile(`
-      <script setup lang="ts">
-      export type Props = { x?: number }
-      defineProps<Props>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`x: { type: Number, required: false }`)
-      expect(bindings).toStrictEqual({
-        x: BindingTypes.PROPS
-      })
-    })
-
-    test('defineProps w/ TS assertion', () => {
-      const { content, bindings } = compile(`
-      <script setup lang="ts">
-        defineProps(['foo'])! as any
-      </script>
-    `)
-      expect(content).toMatch(`props: ['foo']`)
-      assertCode(content)
-      expect(bindings).toStrictEqual({
-        foo: BindingTypes.PROPS
-      })
-    })
-
-    test('withDefaults (static)', () => {
-      const { content, bindings } = compile(`
-      <script setup lang="ts">
-      const props = withDefaults(defineProps<{
-        foo?: string
-        bar?: number;
-        baz: boolean;
-        qux?(): number;
-        quux?(): void
-        quuxx?: Promise<string>;
-        fred?: string
-      }>(), {
-        foo: 'hi',
-        qux() { return 1 },
-        ['quux']() { },
-        async quuxx() { return await Promise.resolve('hi') },
-        get fred() { return 'fred' }
-      })
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(
-        `foo: { type: String, required: false, default: 'hi' }`
-      )
-      expect(content).toMatch(`bar: { type: Number, required: false }`)
-      expect(content).toMatch(`baz: { type: Boolean, required: true }`)
-      expect(content).toMatch(
-        `qux: { type: Function, required: false, default() { return 1 } }`
-      )
-      expect(content).toMatch(
-        `quux: { type: Function, required: false, default() { } }`
-      )
-      expect(content).toMatch(
-        `quuxx: { type: Promise, required: false, async default() { return await Promise.resolve('hi') } }`
-      )
-      expect(content).toMatch(
-        `fred: { type: String, required: false, get default() { return 'fred' } }`
-      )
-      expect(content).toMatch(`const props = __props`)
-      expect(bindings).toStrictEqual({
-        foo: BindingTypes.PROPS,
-        bar: BindingTypes.PROPS,
-        baz: BindingTypes.PROPS,
-        qux: BindingTypes.PROPS,
-        quux: BindingTypes.PROPS,
-        quuxx: BindingTypes.PROPS,
-        fred: BindingTypes.PROPS,
-        props: BindingTypes.SETUP_CONST
-      })
-    })
-
-    test('withDefaults (static) + normal script', () => {
-      const { content } = compile(`
-      <script lang="ts">
-        interface Props {
-          a?: string;
-        }
-      </script>
-      <script setup lang="ts">
-        const props = withDefaults(defineProps<Props>(), {
-          a: "a",
-        });
-      </script>
-      `)
-      assertCode(content)
-    })
-
-    // #7111
-    test('withDefaults (static) w/ production mode', () => {
-      const { content } = compile(
-        `
-      <script setup lang="ts">
-      const props = withDefaults(defineProps<{
-        foo: () => void
-        bar: boolean
-        baz: boolean | (() => void)
-        qux: string | number
-      }>(), {
-        baz: true,
-        qux: 'hi'
-      })
-      </script>
-      `,
-        { isProd: true }
-      )
-      assertCode(content)
-      expect(content).toMatch(`const props = __props`)
-
-      // foo has no default value, the Function can be dropped
-      expect(content).toMatch(`foo: {}`)
-      expect(content).toMatch(`bar: { type: Boolean }`)
-      expect(content).toMatch(
-        `baz: { type: [Boolean, Function], default: true }`
-      )
-      expect(content).toMatch(`qux: { default: 'hi' }`)
-    })
-
-    test('withDefaults (dynamic)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      import { defaults } from './foo'
-      const props = withDefaults(defineProps<{
-        foo?: string
-        bar?: number
-        baz: boolean
-      }>(), { ...defaults })
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`import { mergeDefaults as _mergeDefaults`)
-      expect(content).toMatch(
-        `
-  _mergeDefaults({
-    foo: { type: String, required: false },
-    bar: { type: Number, required: false },
-    baz: { type: Boolean, required: true }
-  }, { ...defaults })`.trim()
-      )
-    })
-
-    test('withDefaults (reference)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      import { defaults } from './foo'
-      const props = withDefaults(defineProps<{
-        foo?: string
-        bar?: number
-        baz: boolean
-      }>(), defaults)
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`import { mergeDefaults as _mergeDefaults`)
-      expect(content).toMatch(
-        `
-  _mergeDefaults({
-    foo: { type: String, required: false },
-    bar: { type: Number, required: false },
-    baz: { type: Boolean, required: true }
-  }, defaults)`.trim()
-      )
-    })
-
-    // #7111
-    test('withDefaults (dynamic) w/ production mode', () => {
-      const { content } = compile(
-        `
-      <script setup lang="ts">
-      import { defaults } from './foo'
-      const props = withDefaults(defineProps<{
-        foo: () => void
-        bar: boolean
-        baz: boolean | (() => void)
-        qux: string | number
-      }>(), { ...defaults })
-      </script>
-      `,
-        { isProd: true }
-      )
-      assertCode(content)
-      expect(content).toMatch(`import { mergeDefaults as _mergeDefaults`)
-      expect(content).toMatch(
-        `
-  _mergeDefaults({
-    foo: { type: Function },
-    bar: { type: Boolean },
-    baz: { type: [Boolean, Function] },
-    qux: {}
-  }, { ...defaults })`.trim()
-      )
-    })
-
-    test('withDefaults w/ dynamic object method', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      const props = withDefaults(defineProps<{
-        foo?: () => 'string'
-      }>(), {
-        ['fo' + 'o']() { return 'foo' }
-      })
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`import { mergeDefaults as _mergeDefaults`)
-      expect(content).toMatch(
-        `
-  _mergeDefaults({
-    foo: { type: Function, required: false }
-  }, {
-        ['fo' + 'o']() { return 'foo' }
-      })`.trim()
-      )
-    })
-
-    test('defineEmits w/ type', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      const emit = defineEmits<(e: 'foo' | 'bar') => void>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-    })
-
-    test('defineEmits w/ type (union)', () => {
-      const type = `((e: 'foo' | 'bar') => void) | ((e: 'baz', id: number) => void)`
-      expect(() =>
-        compile(`
-      <script setup lang="ts">
-      const emit = defineEmits<${type}>()
-      </script>
-      `)
-      ).toThrow()
-    })
-
-    test('defineEmits w/ type (type literal w/ call signatures)', () => {
-      const type = `{(e: 'foo' | 'bar'): void; (e: 'baz', id: number): void;}`
-      const { content } = compile(`
-      <script setup lang="ts">
-      const emit = defineEmits<${type}>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar", "baz"]`)
-    })
-
-    test('defineEmits w/ type (interface)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      interface Emits { (e: 'foo' | 'bar'): void }
-      const emit = defineEmits<Emits>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-    })
-
-    test('defineEmits w/ type (exported interface)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      export interface Emits { (e: 'foo' | 'bar'): void }
-      const emit = defineEmits<Emits>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-    })
-
-    test('defineEmits w/ type from normal script', () => {
-      const { content } = compile(`
-      <script lang="ts">
-        export interface Emits { (e: 'foo' | 'bar'): void }
-      </script>
-      <script setup lang="ts">
-      const emit = defineEmits<Emits>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-    })
-
-    test('defineEmits w/ type (type alias)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      type Emits = { (e: 'foo' | 'bar'): void }
-      const emit = defineEmits<Emits>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-    })
-
-    test('defineEmits w/ type (exported type alias)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      export type Emits = { (e: 'foo' | 'bar'): void }
-      const emit = defineEmits<Emits>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-    })
-
-    test('defineEmits w/ type (referenced function type)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      type Emits = (e: 'foo' | 'bar') => void
-      const emit = defineEmits<Emits>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-    })
-
-    test('defineEmits w/ type (referenced exported function type)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      export type Emits = (e: 'foo' | 'bar') => void
-      const emit = defineEmits<Emits>()
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-    })
-
-    // #5393
-    test('defineEmits w/ type (interface ts type)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      interface Emits { (e: 'foo'): void }
-      const emit: Emits = defineEmits(['foo'])
-      </script>
-      `)
-      assertCode(content)
-      expect(content).toMatch(`emits: ['foo']`)
-    })
-
-    test('defineEmits w/ type (property syntax)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      const emit = defineEmits<{ foo: [], bar: [] }>()
-      </script>
-      `)
-      expect(content).toMatch(`emits: ["foo", "bar"]`)
-      assertCode(content)
-    })
-
-    // #8040
-    test('defineEmits w/ type (property syntax string literal)', () => {
-      const { content } = compile(`
-      <script setup lang="ts">
-      const emit = defineEmits<{ 'foo:bar': [] }>()
-      </script>
-      `)
-      expect(content).toMatch(`emits: ["foo:bar"]`)
-      assertCode(content)
-    })
-
-    describe('defineSlots()', () => {
-      test('basic usage', () => {
-        const { content } = compile(`
-          <script setup lang="ts">
-          const slots = defineSlots<{
-            default: { msg: string }
-          }>()
-          </script>
-        `)
-        assertCode(content)
-        expect(content).toMatch(`const slots = _useSlots()`)
-        expect(content).not.toMatch('defineSlots')
-      })
-
-      test('w/o return value', () => {
-        const { content } = compile(`
-          <script setup lang="ts">
-          defineSlots<{
-            default: { msg: string }
-          }>()
-          </script>
-        `)
-        assertCode(content)
-        expect(content).not.toMatch('defineSlots')
-        expect(content).not.toMatch(`_useSlots`)
-      })
-
-      test('w/o generic params', () => {
-        const { content } = compile(`
-          <script setup>
-          const slots = defineSlots()
-          </script>
-        `)
-        assertCode(content)
-        expect(content).toMatch(`const slots = _useSlots()`)
-        expect(content).not.toMatch('defineSlots')
-      })
-    })
-
-    describe('defineModel()', () => {
-      test('basic usage', () => {
-        const { content, bindings } = compile(
-          `
-          <script setup lang="ts">
-          const modelValue = defineModel<boolean | string>()
-          const count = defineModel<number>('count')
-          const disabled = defineModel<number>('disabled', { required: false })
-          const any = defineModel<any | boolean>('any')
-          </script>
-          `,
-          { defineModel: true }
-        )
-        assertCode(content)
-        expect(content).toMatch('"modelValue": { type: [Boolean, String] }')
-        expect(content).toMatch('"count": { type: Number }')
-        expect(content).toMatch(
-          '"disabled": { type: Number, ...{ required: false } }'
-        )
-        expect(content).toMatch('"any": { type: Boolean, skipCheck: true }')
-        expect(content).toMatch(
-          'emits: ["update:modelValue", "update:count", "update:disabled", "update:any"]'
-        )
-
-        expect(content).toMatch(
-          `const modelValue = _useModel(__props, "modelValue")`
-        )
-        expect(content).toMatch(`const count = _useModel(__props, "count")`)
-        expect(content).toMatch(
-          `const disabled = _useModel(__props, "disabled")`
-        )
-        expect(content).toMatch(`const any = _useModel(__props, "any")`)
-
-        expect(bindings).toStrictEqual({
-          modelValue: BindingTypes.SETUP_REF,
-          count: BindingTypes.SETUP_REF,
-          disabled: BindingTypes.SETUP_REF,
-          any: BindingTypes.SETUP_REF
-        })
-      })
-
-      test('w/ production mode', () => {
-        const { content, bindings } = compile(
-          `
-          <script setup lang="ts">
-          const modelValue = defineModel<boolean>()
-          const fn = defineModel<() => void>('fn')
-          const fnWithDefault = defineModel<() => void>('fnWithDefault', { default: () => null })
-          const str = defineModel<string>('str')
-          const optional = defineModel<string>('optional', { required: false })
-          </script>
-          `,
-          { defineModel: true, isProd: true }
-        )
-        assertCode(content)
-        expect(content).toMatch('"modelValue": { type: Boolean }')
-        expect(content).toMatch('"fn": {}')
-        expect(content).toMatch(
-          '"fnWithDefault": { type: Function, ...{ default: () => null } },'
-        )
-        expect(content).toMatch('"str": {}')
-        expect(content).toMatch('"optional": { required: false }')
-        expect(content).toMatch(
-          'emits: ["update:modelValue", "update:fn", "update:fnWithDefault", "update:str", "update:optional"]'
-        )
-        expect(content).toMatch(
-          `const modelValue = _useModel(__props, "modelValue")`
-        )
-        expect(content).toMatch(`const fn = _useModel(__props, "fn")`)
-        expect(content).toMatch(`const str = _useModel(__props, "str")`)
-        expect(bindings).toStrictEqual({
-          modelValue: BindingTypes.SETUP_REF,
-          fn: BindingTypes.SETUP_REF,
-          fnWithDefault: BindingTypes.SETUP_REF,
-          str: BindingTypes.SETUP_REF,
-          optional: BindingTypes.SETUP_REF
-        })
-      })
-    })
-
     test('runtime Enum', () => {
       const { content, bindings } = compile(
         `<script setup lang="ts">
@@ -1944,56 +859,6 @@ const emit = defineEmits(['a', 'b'])
       expect(bindings).toStrictEqual({
         Foo: BindingTypes.LITERAL_CONST
       })
-    })
-
-    test('runtime inference for Enum in defineProps', () => {
-      expect(
-        compile(
-          `<script setup lang="ts">
-        const enum Foo { A = 123 }
-        defineProps<{
-          foo: Foo
-        }>()
-        </script>`,
-          { hoistStatic: true }
-        ).content
-      ).toMatch(`foo: { type: Number`)
-
-      expect(
-        compile(
-          `<script setup lang="ts">
-        const enum Foo { A = '123' }
-        defineProps<{
-          foo: Foo
-        }>()
-        </script>`,
-          { hoistStatic: true }
-        ).content
-      ).toMatch(`foo: { type: String`)
-
-      expect(
-        compile(
-          `<script setup lang="ts">
-        const enum Foo { A = '123', B = 123 }
-        defineProps<{
-          foo: Foo
-        }>()
-        </script>`,
-          { hoistStatic: true }
-        ).content
-      ).toMatch(`foo: { type: [String, Number]`)
-
-      expect(
-        compile(
-          `<script setup lang="ts">
-        const enum Foo { A, B }
-        defineProps<{
-          foo: Foo
-        }>()
-        </script>`,
-          { hoistStatic: true }
-        ).content
-      ).toMatch(`foo: { type: Number`)
     })
 
     test('import type', () => {
@@ -2154,20 +1019,6 @@ const emit = defineEmits(['a', 'b'])
       ).toThrow(moduleErrorMsg)
     })
 
-    test('defineProps/Emit() w/ both type and non-type args', () => {
-      expect(() => {
-        compile(`<script setup lang="ts">
-        defineProps<{}>({})
-        </script>`)
-      }).toThrow(`cannot accept both type and non-type arguments`)
-
-      expect(() => {
-        compile(`<script setup lang="ts">
-        defineEmits<{}>({})
-        </script>`)
-      }).toThrow(`cannot accept both type and non-type arguments`)
-    })
-
     test('defineProps/Emit() referencing local var', () => {
       expect(() =>
         compile(`<script setup>
@@ -2230,19 +1081,6 @@ const emit = defineEmits(['a', 'b'])
           foo: () => bar > 1
         })
         </script>`).content
-      )
-    })
-
-    test('mixed usage of property / call signature in defineEmits', () => {
-      expect(() =>
-        compile(`<script setup lang="ts">
-        defineEmits<{
-          foo: []
-          (e: 'hi'): void
-        }>()
-        </script>`)
-      ).toThrow(
-        `defineEmits() type cannot mixed call signature and property syntax.`
       )
     })
   })
